requirements.txt
flask==2.3.2
pandas==1.5.3
matplotlib==3.7.1
seaborn==0.12.2
scikit-learn==1.2.2
google-cloud-storage==2.11.0
requests==2.31.0
apscheduler==3.10.1
pytz==2023.3